clean-room-demo/webapp/backend/requirements.txt
starlette==0.14.2
databases==0.4.3
aiofiles==0.6.0
jinja2==2.11.3
python-dotenv==0.17.0
aiohttp==3.7.4.post0
uvicorn==0.13.4
uvloop==0.15.2
#TODO: Required in docker?
httptools==0.2.0
dataclasses_serialization==1.3.1
azure-confidentialledger==1.0.0b1
azure-mgmt-confidentialledger==1.0.0b1
azure-identity==1.6.0
requests==2.26.0
SQLAlchem==1.4.23
yubico-client==1.13.0